+++ TensorWeaver/OutputData/Classification.cs
@@ -2,10 +2,10 @@ namespace TensorWeaver.OutputData;
 
 public readonly struct Classification
 {
-	public ushort ClassId { get; }
+	public byte ClassId { get; }
 	public float Confidence { get; }
 
-	public Classification(ushort classId, float confidence)
+	public Classification(byte classId, float confidence)
 	{
 		ClassId = classId;
 		Confidence = confidence;
